Expose repo intelligence tool surface
