+++ crates/tandem-server/src/app/state/automation/logic.rs
@@ -2,6 +2,7 @@ use super::*;
 
 include!("logic_parts/part01.rs");
 include!("logic_parts/part02.rs");
+include!("logic_parts/part07.rs");
 include!("logic_parts/part03.rs");
 include!("logic_parts/part04.rs");
 include!("logic_parts/part06.rs");
