+++ crates/tandem-server/src/app/state/automation/mod.rs
@@ -123,4 +123,6 @@ mod logic;
 pub(crate) use logic::*;
 
 #[cfg(test)]
+mod repo_context_policy_tests;
+#[cfg(test)]
 mod tests;
